% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-db.R
\name{connect_khelsa}
\alias{connect_khelsa}
\title{Connect to Database}
\usage{
connect_khelsa(db = c("kh", "logg"), dbname = NULL, .test = FALSE, ...)
}
\arguments{
\item{db}{Database file \code{kh} (KHELSA) or \code{logg} (KHlogg)}

\item{dbname}{Database filename with complete path}

\item{.test}{Use for testing only, returns db name}

\item{...}{Other arguments}
}
\description{
Produce an R6 object for connecting to the database
}
\keyword{internal}
