% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{do_cube_params}
\alias{do_cube_params}
\title{do_cube_params}
\usage{
do_cube_params(
  con_db,
  con_logg,
  con_duck,
  cube.id = cube.id,
  batch = batch,
  year = year
)
}
\arguments{
\item{con_db}{connection to KHELSA}

\item{con_logg}{connection to KHlogg}

\item{con_duck}{connection to in-memory duck db}

\item{cube.id}{name of cube}

\item{batch}{batch date}

\item{year}{production year}
}
\value{
list containing all relevant information for cube production
including formatted filegroups.
}
\description{
Collects all relevant parameters and files necessary to produce the finished
data file. Reads tables from access, reads and filters file groups.
}
\concept{worker}
\keyword{internal}
