% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{do_readfilegroup}
\alias{do_readfilegroup}
\title{do_readfile}
\usage{
do_readfilegroup(filepath, tabfilter, filter)
}
\arguments{
\item{filepath}{path to file}

\item{tabfilter}{string for filtering TAB-columns}

\item{filter}{TRUE/FALSE based on is_filter()}
}
\value{
filtered data.table
}
\description{
do_readfile
}
\keyword{internal}
