% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{find_files}
\alias{find_files}
\title{find_files}
\usage{
find_files(KUBER, TNP_PROD, STNP_PROD)
}
\arguments{
\item{KUBER}{ACCESS table}

\item{TNP_PROD}{ACCESS table}

\item{STNP_PROD}{ACCESS table}
}
\value{
a list containing the names of files needed to create the final data
}
\description{
find_files
}
\seealso{
Other helper: 
\code{\link{get_access_cube}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_filter}()},
\code{\link{is_teller}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
