% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{get_access_cube}
\alias{get_access_cube}
\title{get_access_cube}
\usage{
get_access_cube(params = params)
}
\arguments{
\item{params}{list of params}
}
\value{
updated params list
}
\description{
Generate a list of necessary information to generate the data file.
The function reads tables in KHELSA Access database and returns the relevant
rows, and a list of files to be used to create the finished cube.
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_filter}()},
\code{\link{is_teller}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
