% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{get_design_cube}
\alias{get_design_cube}
\title{get_design_cube}
\usage{
get_design_cube(params)
}
\arguments{
\item{params}{a list of parameters}
}
\value{
updated params list including DESIGN parameters
}
\description{
Fetches and format design parameters from ACCESS
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_access_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_filter}()},
\code{\link{is_teller}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
