% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{get_filegroup_path}
\alias{get_filegroup_path}
\title{is_filegroup_path}
\usage{
get_filegroup_path(filename)
}
\arguments{
\item{filename}{name of filegroup}
}
\value{
Path to filegroup
}
\description{
Fetch the path to the most recent file corresponding to the filegroup name
from the filegroup folder.
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_access_cube}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_filter}()},
\code{\link{is_teller}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
