% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{get_filegroupdesign}
\alias{get_filegroupdesign}
\title{do_filegroupdesign}
\usage{
get_filegroupdesign(files, params)
}
\arguments{
\item{params}{}
}
\value{
design of filegroup
}
\description{
do_filegroupdesign
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_access_cube}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_filter}()},
\code{\link{is_teller}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
