% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{get_tabsubset}
\alias{get_tabsubset}
\title{get_tabsubset}
\usage{
get_tabsubset(filename, params = params)
}
\arguments{
\item{filename}{name of filegroup}

\item{params}{List of access params generated with \code{ìs_access_cube}}
}
\value{
formatted filtering string for TAB-columns.
}
\description{
Check for filters set up in ACCESS::KUBER::TAB-columns and set up a filter to
remove unneeded categories. Uses values in \code{TABx} unless \code{TABx_0} is non-empty,
then values in \code{TABx_0}is used. Only applies to \code{TELLER}-file.
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_access_cube}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{is_filter}()},
\code{\link{is_teller}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
