% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_coltype}
\alias{is_coltype}
\title{is_coltype}
\usage{
is_coltype(dt)
}
\arguments{
\item{dt}{data table}
}
\value{
list of values and dimensions
}
\description{
is_coltype
}
\seealso{
Other utils: 
\code{\link{is_fileaccess}()},
\code{\link{is_orgcube_path}()},
\code{\link{update_orgcube}()}
}
\concept{utils}
\keyword{internal}
