% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_fileaccess}
\alias{is_fileaccess}
\title{is_fileaccess}
\usage{
is_fileaccess(file)
}
\description{
Uses file.access() to check if file exists and can be read
}
\seealso{
Other utils: 
\code{\link{is_coltype}()},
\code{\link{is_orgcube_path}()},
\code{\link{update_orgcube}()}
}
\concept{utils}
\keyword{internal}
