% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{is_filter}
\alias{is_filter}
\title{is_filter}
\usage{
is_filter(filegroup, params)
}
\arguments{
\item{filegroup}{name of filegroup}

\item{params}{access params}
}
\description{
Check if filters are set up in ACCESS::FILFILTRE
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_access_cube}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_teller}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
