% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{is_orgcube_path}
\alias{is_orgcube_path}
\title{is_orgcube_path}
\usage{
is_orgcube_path(
  folder = c("kh", "nh", "data", "dumps", "filegroups"),
  type = NULL,
  format = NULL
)
}
\arguments{
\item{folder}{one of c("kh", "nh", "data", "dumps", "filegroups")}

\item{type}{NULL or one of c("datert", "nyeste", "qc", "specs")}

\item{format}{NULL or c("R", "csv")}
}
\value{
desired path. Will throw error if path does not exist.
}
\description{
is_orgcube_path
}
\seealso{
Other utils: 
\code{\link{is_coltype}()},
\code{\link{is_fileaccess}()},
\code{\link{update_orgcube}()}
}
\concept{utils}
\keyword{internal}
