% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{is_teller}
\alias{is_teller}
\title{is_teller}
\usage{
is_teller(filename, params)
}
\arguments{
\item{params}{access params}

\item{filegroup}{name of filegroup}
}
\description{
Check if file is equal to the TELLER file
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_access_cube}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_filter}()},
\code{\link{use_filegroup_filter}()}
}
\concept{helper}
\keyword{internal}
