% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{make_cube}
\alias{make_cube}
\title{Make Cube}
\usage{
make_cube(cube.id = NULL, year = NULL)
}
\arguments{
\item{cube.id}{name of cube, corresponding to an entry in KHELSA::KUBER::KUBE_NAVN}

\item{year}{production year}
}
\value{
cube
}
\description{
Performs all data processing starting from a filegroup generated with
\code{\link[=make_filegroup]{make_filegroup()}} and returns a complete publication-ready data file.
}
\concept{main}
