% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filegroup.R
\name{make_filegroup}
\alias{make_filegroup}
\title{Make Filegroup}
\usage{
make_filegroup(group.id = NULL)
}
\arguments{
\item{group.id}{name of filegroup, corresponding to an entry in KHELSA::FILGRUPPER::FILGRUPPE}
}
\value{
a complete filegroup
}
\description{
Reads and process original data and returns a filegroup which can be
further processed by \code{\link[=make_cube]{make_cube()}}

Performs all data processing starting from a filegroup and returns a
complete publication-ready data file.
}
