% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgcube-package.R
\docType{package}
\name{orgcube-package}
\alias{orgcube}
\alias{orgcube-package}
\title{orgcube: Data processing for public health profiles}
\description{
Aggregate, process, and formats data for publication
}
\author{
\strong{Maintainer}: Vegard Lysne \email{vegard.lysne@helsedir.no} (\href{https://orcid.org/0000-0002-0816-5075}{ORCID})

}
\keyword{internal}
