% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-files.R
\name{read_files}
\alias{read_files}
\title{Read Files}
\usage{
read_files(specs = specs)
}
\arguments{
\item{specs}{List of access specs generated with \code{\link[=is_access_cube]{is_access_cube()}}}
}
\value{
description
}
\description{
Uses information from ACCESS to read and filter the original files
to be used to create the final data using \code{\link[=make_cube]{make_cube()}}. Returns a list
of formatted files.
}
