% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{update_orgcube}
\alias{update_orgcube}
\title{Update package}
\usage{
update_orgcube(branch = "main", force = FALSE)
}
\arguments{
\item{branch}{The branch in Github to install from, \code{main} or \code{dev}}

\item{force}{Use the laterst version(s) of all dependencies. Default is FALSE
\dontrun{
update_orgcube(branch = "main") #default
update_orgcube(branch = "dev") #to upgrade with dev branch
}}
}
\description{
Update orgcube directly with Github repo. Default is \code{main} branch.
}
\seealso{
Other utils: 
\code{\link{is_coltype}()},
\code{\link{is_fileaccess}()},
\code{\link{is_orgcube_path}()}
}
\concept{utils}
