% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube-params.R
\name{use_filegroup_filter}
\alias{use_filegroup_filter}
\title{use_filegroup_filter}
\usage{
use_filegroup_filter(dt, filterspec, params)
}
\arguments{
\item{file}{a data.table}

\item{filerparams}{relevant params from KHELSA:FILFILTRE}
}
\value{
a filtered data file
}
\description{
Filters a data file according to KHELSA:FILFILTRE
}
\seealso{
Other helper: 
\code{\link{find_files}()},
\code{\link{get_access_cube}()},
\code{\link{get_design_cube}()},
\code{\link{get_filegroup_path}()},
\code{\link{get_filegroupdesign}()},
\code{\link{get_filegroups}()},
\code{\link{get_tabsubset}()},
\code{\link{is_filter}()},
\code{\link{is_teller}()}
}
\concept{helper}
\keyword{internal}
